+++ packages/ui/src/lib/input/input.component.knobs.ts
@@ -32,6 +32,7 @@ export const INPUT_KNOBS: ComponentKnobs = {
     autofocus: { control: 'boolean' },
     showPasswordToggle: { control: 'boolean', if: { arg: 'type', eq: 'password' } },
     clearable: { control: 'boolean' },
+    keepIcon: { control: 'boolean' },
     triggerError: { control: 'boolean', demoOnly: true },
     focused: { action: 'focused' },
     blurred: { action: 'blurred' },
@@ -47,6 +48,7 @@ export const INPUT_KNOBS: ComponentKnobs = {
     autofocus: false,
     showPasswordToggle: true,
     clearable: false,
+    keepIcon: true,
     triggerError: false,
   },
 };